@@ -1,8 +1,7 @@
 import React from 'react';
 import sinon from 'sinon';
 import { mount } from 'enzyme';
-
-import Formsy, { withFormsy } from '..';
+import Formsy, { withFormsy } from '../src';
 import immediate from '../__test_utils__/immediate';
 import TestInput, { InputFactory } from '../__test_utils__/TestInput';
 
@@ -49,7 +48,7 @@ describe('Element', () => {
   it('should set back to pristine value when running reset', () => {
     let reset = null;
     const Input = InputFactory({
-      componentDidMount() {
+      componentDidUpdate() {
         reset = this.props.resetValue;
       },
     });
@@ -68,7 +67,7 @@ describe('Element', () => {
   it('should return error message passed when calling getErrorMessage()', () => {
     let errorMessage = null;
     const Input = InputFactory({
-      componentDidMount() {
+      componentDidUpdate() {
         errorMessage = this.props.errorMessage;
       },
     });
@@ -84,8 +83,8 @@ describe('Element', () => {
   it('should return true or false when calling isValid() depending on valid state', () => {
     let isValid = null;
     const Input = InputFactory({
-      componentWillReceiveProps: function(nextProps) {
-        isValid = nextProps.isValid;
+      componentDidUpdate: function() {
+        isValid = this.props.isValid;
       },
     });
     const form = mount(
@@ -103,7 +102,7 @@ describe('Element', () => {
   it('should return true or false when calling isRequired() depending on passed required attribute', () => {
     const isRequireds = [];
     const Input = InputFactory({
-      componentDidMount() {
+      componentDidUpdate() {
         isRequireds.push(this.props.isRequired);
       },
     });
@@ -123,7 +122,7 @@ describe('Element', () => {
   it('should return true or false when calling showRequired() depending on input being empty and required is passed, or not', () => {
     const showRequireds = [];
     const Input = InputFactory({
-      componentDidMount() {
+      componentDidUpdate() {
         showRequireds.push(this.props.showRequired);
       },
     });
@@ -607,5 +606,18 @@ describe('Element', () => {
         <TestInput ref={onInputRef} name="name" value="foo" />
       </Formsy>,
     );
+  });
+
+  it('unregisters on unmount', () => {
+    const TestComponent = ({ hasInput }) => <Formsy>{hasInput ? <TestInput name="foo" value="foo" /> : null}</Formsy>;
+
+    const wrapper = mount(<TestComponent hasInput />);
+    const formsy = wrapper.find(Formsy).instance();
+
+    expect(formsy.inputs).toHaveLength(1);
+
+    wrapper.setProps({ hasInput: false });
+
+    expect(formsy.inputs).toHaveLength(0);
   });
 });
