@@ -1,11 +1,10 @@
-import React from 'react';
 import { mount } from 'enzyme';
-import sinon from 'sinon';
-
-import Formsy, { addValidationRule } from './..';
+import * as React from 'react';
+import DynamicInputForm from '../__test_utils__/DynamicInputForm';
 import TestInput from '../__test_utils__/TestInput';
 import TestInputHoc from '../__test_utils__/TestInputHoc';
-import immediate from '../__test_utils__/immediate';
+
+import Formsy, { addValidationRule } from '../src';
 
 describe('Setting up a form', () => {
   it('should expose the users DOM node through an innerRef prop', () => {
@@ -41,6 +40,7 @@ describe('Setting up a form', () => {
 
   it('should allow for null/undefined children', () => {
     let model = null;
+
     class TestForm extends React.Component {
       render() {
         return (
@@ -48,112 +48,82 @@ describe('Setting up a form', () => {
             <h1>Test</h1>
             {null}
             {undefined}
-            <TestInput name="name" value={'foo'} />
+            <TestInput name="name" value="foo" />
           </Formsy>
         );
       }
     }
 
     const form = mount(<TestForm />);
-    immediate(() => {
-      expect(model).toEqual({ name: 'foo' });
-    });
+    form.simulate('submit');
+    expect(model).toEqual({ name: 'foo' });
   });
 
   it('should allow for inputs being added dynamically', () => {
-    const inputs = [];
-    let forceUpdate = null;
     let model = null;
-    class TestForm extends React.Component {
-      componentWillMount() {
-        forceUpdate = this.forceUpdate.bind(this);
-      }
-      render() {
-        return <Formsy onSubmit={formModel => (model = formModel)}>{inputs}</Formsy>;
-      }
-    }
-    const form = mount(<TestForm />);
 
-    // Wait before adding the input
-    setTimeout(() => {
-      inputs.push(<TestInput name="test" value="" key={inputs.length} />);
+    const form = mount(<DynamicInputForm onSubmit={formModel => (model = formModel)} inputName="test" />);
+    form.find('button').simulate('click');
+    form.update();
 
-      forceUpdate(() => {
-        // Wait for next event loop, as that does the form
-        immediate(() => {
-          form.simulate('submit');
-          test.ok('test' in model);
-        });
-      });
-    }, 10);
+    form.simulate('submit');
+    expect(model).toHaveProperty('test');
   });
 
   it('should allow dynamically added inputs to update the form-model', () => {
-    const inputs = [];
-    let forceUpdate = null;
     let model = null;
-    class TestForm extends React.Component {
-      componentWillMount() {
-        forceUpdate = this.forceUpdate.bind(this);
-      }
-      render() {
-        return <Formsy onSubmit={formModel => (model = formModel)}>{inputs}</Formsy>;
-      }
-    }
-    const form = mount(<TestForm />);
 
-    // Wait before adding the input
-    immediate(() => {
-      inputs.push(<TestInput name="test" key={inputs.length} />);
+    const form = mount(<DynamicInputForm onSubmit={formModel => (model = formModel)} inputName="test" />);
+    form.find('button').simulate('click');
+    form.update();
 
-      forceUpdate(() => {
-        // Wait for next event loop, as that does the form
-        immediate(() => {
-          form.find('input').simulate({
-            target: { value: 'foo' },
-          });
-          form.simulate('submit');
-          expect(model.test).toEqual('foo');
-        });
-      });
+    form.find('input').simulate('change', {
+      target: { value: 'foo' },
     });
+    form.simulate('submit');
+    expect(model).toHaveProperty('test', 'foo');
   });
 
   it('should allow a dynamically updated input to update the form-model', () => {
-    let forceUpdate = null;
     let model = null;
 
-    class TestForm extends React.Component {
-      componentWillMount() {
-        forceUpdate = this.forceUpdate.bind(this);
-      }
+    class TestForm extends React.Component<any> {
+      public state = {
+        inputValue: this.props.inputValue,
+      };
+
       render() {
-        const input = <TestInput name="test" value={this.props.value} />;
-
-        return <Formsy onSubmit={formModel => (model = formModel)}>{input}</Formsy>;
+        const { inputValue } = this.state;
+        return (
+          <Formsy onSubmit={formModel => (model = formModel)}>
+            <TestInput name="test" value={inputValue} />
+            <button type="button" onClick={this.updateInputValue} />
+          </Formsy>
+        );
       }
+
+      updateInputValue = () => this.setState({ inputValue: 'bar' });
     }
-    let form = mount(<TestForm value="foo" />);
 
-    // Wait before changing the input
-    immediate(() => {
-      form = mount(<TestForm value="bar" />);
+    const form = mount(<TestForm inputValue="foo" />);
 
-      forceUpdate(() => {
-        // Wait for next event loop, as that does the form
-        immediate(() => {
-          form.simulate('submit');
-          expect(model.test).toEqual('bar');
-        });
-      });
-    });
+    form.simulate('submit');
+
+    expect(model).toHaveProperty('test', 'foo');
+
+    form.find('button').simulate('click');
+    form.update();
+    form.simulate('submit');
+    form.update();
+
+    expect(model).toHaveProperty('test', 'bar');
   });
 });
 
 describe('validations', () => {
   it('should run when the input changes', () => {
-    const runRule = sinon.spy();
-    const notRunRule = sinon.spy();
+    const runRule = jest.fn();
+    const notRunRule = jest.fn();
 
     addValidationRule('runRule', runRule);
     addValidationRule('notRunRule', notRunRule);
@@ -168,13 +138,14 @@ describe('validations', () => {
     input.simulate('change', {
       target: { value: 'bar' },
     });
-    expect(runRule.calledWith({ one: 'bar' })).toBe(true);
-    expect(notRunRule.called).toEqual(false);
+
+    expect(runRule).toHaveBeenCalledWith({ one: 'bar' }, 'bar', true);
+    expect(notRunRule).not.toHaveBeenCalled();
   });
 
   it('should allow the validation to be changed', () => {
-    const ruleA = sinon.spy();
-    const ruleB = sinon.spy();
+    const ruleA = jest.fn();
+    const ruleB = jest.fn();
     addValidationRule('ruleA', ruleA);
     addValidationRule('ruleB', ruleB);
 
@@ -183,11 +154,13 @@ describe('validations', () => {
         super(props);
         this.state = { rule: 'ruleA' };
       }
+
       changeRule = () => {
         this.setState({
           rule: 'ruleB',
         });
       };
+
       render() {
         return (
           <Formsy>
@@ -203,25 +176,29 @@ describe('validations', () => {
     input.simulate('change', {
       target: { value: 'bar' },
     });
-    expect(ruleB.calledWith({ one: 'bar' })).toBe(true);
+    expect(ruleB).toHaveBeenCalledWith({ one: 'bar' }, 'bar', true);
   });
 
   it('should invalidate a form if dynamically inserted input is invalid', () => {
-    const isInValidSpy = sinon.spy();
+    const isInValidSpy = jest.fn();
 
     class TestForm extends React.Component {
+      formRef = React.createRef();
+
       constructor(props) {
         super(props);
         this.state = { showSecondInput: false };
       }
+
       addInput = () => {
         this.setState({
           showSecondInput: true,
         });
       };
+
       render() {
         return (
-          <Formsy ref="formsy" onInvalid={isInValidSpy}>
+          <Formsy ref={this.formRef} onInvalid={isInValidSpy}>
             <TestInput name="one" validations="isEmail" value="foo@bar.com" />
             {this.state.showSecondInput ? <TestInput name="two" validations="isEmail" value="foo@bar" /> : null}
           </Formsy>
@@ -231,30 +208,32 @@ describe('validations', () => {
 
     const form = mount(<TestForm />);
 
-    expect(form.instance().refs.formsy.state.isValid).toEqual(true);
+    expect(form.instance().formRef.current.state.isValid).toEqual(true);
     form.instance().addInput();
 
-    immediate(() => {
-      expect(isInValidSpy.called).toEqual(true);
-    });
+    expect(isInValidSpy).toHaveBeenCalled();
   });
 
   it('should validate a form when removing an invalid input', () => {
-    const isValidSpy = sinon.spy();
+    const isValidSpy = jest.fn();
 
     class TestForm extends React.Component {
+      formRef = React.createRef();
+
       constructor(props) {
         super(props);
         this.state = { showSecondInput: true };
       }
+
       removeInput() {
         this.setState({
           showSecondInput: false,
         });
       }
+
       render() {
         return (
-          <Formsy ref="formsy" onValid={isValidSpy}>
+          <Formsy ref={this.formRef} onValid={isValidSpy}>
             <TestInput name="one" validations="isEmail" value="foo@bar.com" />
             {this.state.showSecondInput ? <TestInput name="two" validations="isEmail" value="foo@bar" /> : null}
           </Formsy>
@@ -264,17 +243,15 @@ describe('validations', () => {
 
     const form = mount(<TestForm />);
 
-    expect(form.instance().refs.formsy.state.isValid).toEqual(false);
+    expect(form.instance().formRef.current.state.isValid).toEqual(false);
     form.instance().removeInput();
 
-    immediate(() => {
-      expect(isValidSpy.called).toEqual(true);
-    });
+    expect(isValidSpy).toHaveBeenCalled();
   });
 
   it('runs multiple validations', () => {
-    const ruleA = sinon.spy();
-    const ruleB = sinon.spy();
+    const ruleA = jest.fn();
+    const ruleB = jest.fn();
     addValidationRule('ruleA', ruleA);
     addValidationRule('ruleB', ruleB);
 
@@ -288,45 +265,50 @@ describe('validations', () => {
     input.simulate('change', {
       target: { value: 'bar' },
     });
-    expect(ruleA.calledWith({ one: 'bar' })).toBe(true);
-    expect(ruleB.calledWith({ one: 'bar' })).toBe(true);
+    expect(ruleA).toHaveBeenCalledWith({ one: 'bar' }, 'bar', true);
+    expect(ruleB).toHaveBeenCalledWith({ one: 'bar' }, 'bar', true);
   });
 });
 
 describe('onChange', () => {
   it('should not trigger onChange when form is mounted', () => {
-    const hasChanged = sinon.spy();
+    const hasChanged = jest.fn();
+
     class TestForm extends React.Component {
       render() {
         return <Formsy onChange={hasChanged}></Formsy>;
       }
     }
+
     mount(<TestForm />);
-    expect(hasChanged.called).toEqual(false);
+    expect(hasChanged).not.toHaveBeenCalled();
   });
 
   it('should trigger onChange once when form element is changed', () => {
-    const hasChanged = sinon.spy();
+    const hasChanged = jest.fn();
     const form = mount(
       <Formsy onChange={hasChanged}>
-        <TestInput name="foo" />
+        <TestInput name="foo" value="" />
       </Formsy>,
     );
     form.find('input').simulate('change', { target: { value: 'bar' } });
-    expect(hasChanged.calledOnce).toEqual(true);
+    expect(hasChanged).toHaveBeenCalledTimes(1);
   });
 
   it('should trigger onChange once when new input is added to form', () => {
-    const hasChanged = sinon.spy();
+    const hasChanged = jest.fn();
+
     class TestForm extends React.Component {
       state = {
         showInput: false,
       };
+
       addInput() {
         this.setState({
           showInput: true,
         });
       }
+
       render() {
         return <Formsy onChange={hasChanged}>{this.state.showInput ? <TestInput name="test" /> : null}</Formsy>;
       }
@@ -334,22 +316,23 @@ describe('onChange', () => {
 
     const form = mount(<TestForm />);
     form.instance().addInput();
-    immediate(() => {
-      expect(hasChanged.calledOnce).toEqual(true);
-    });
+    expect(hasChanged).toHaveBeenCalledTimes(1);
   });
 
   it('onChange should honor dot notation transformations', () => {
-    const hasChanged = sinon.spy();
+    const hasChanged = jest.fn();
+
     class TestForm extends React.Component {
       state = {
         showInput: false,
       };
+
       addInput() {
         this.setState({
           showInput: true,
         });
       }
+
       render() {
         return (
           <Formsy onChange={hasChanged}>
@@ -361,9 +344,9 @@ describe('onChange', () => {
 
     const form = mount(<TestForm />);
     form.instance().addInput();
-    immediate(() => {
-      expect(hasChanged.args[0][0]).toEqual({ parent: { child: 'test' } });
-    });
+    form.update();
+
+    expect(hasChanged).toHaveBeenCalledWith({ parent: { child: 'test' } }, false);
   });
 });
 
@@ -371,9 +354,11 @@ describe('Update a form', () => {
   it('should allow elements to check if the form is disabled', () => {
     class TestForm extends React.Component {
       state = { disabled: true };
+
       enableForm() {
         this.setState({ disabled: false });
       }
+
       render() {
         return (
           <Formsy disabled={this.state.disabled}>
@@ -388,9 +373,8 @@ describe('Update a form', () => {
     expect(input.instance().isFormDisabled()).toEqual(true);
 
     form.instance().enableForm();
-    immediate(() => {
-      expect(input.instance().isFormDisabled()).toEqual(false);
-    });
+
+    expect(input.instance().isFormDisabled()).toEqual(false);
   });
 
   it('should be possible to pass error state of elements by changing an errors attribute', () => {
@@ -399,31 +383,28 @@ describe('Update a form', () => {
       onChange = values => {
         this.setState(values.foo ? { validationErrors: {} } : { validationErrors: { foo: 'bar' } });
       };
+
       render() {
         return (
           <Formsy onChange={this.onChange} validationErrors={this.state.validationErrors}>
-            <TestInput name="foo" />
+            <TestInput name="foo" value="" />
           </Formsy>
         );
       }
     }
+
     const form = mount(<TestForm />);
 
-    // Wait for update
-    immediate(() => {
-      const input = form.find(TestInput);
-      expect(input.instance().getErrorMessage()).toEqual('bar');
-      input.setValue('gotValue');
+    const input = form.find(TestInput);
+    expect(input.instance().getErrorMessage()).toEqual('bar');
+    input.instance().setValue('gotValue');
 
-      // Wait for update
-      immediate(() => {
-        expect(input.instance().getErrorMessage()).toEqual(null);
-      });
-    });
+    expect(input.instance().getErrorMessage()).toEqual(null);
   });
 
   it('should trigger an onValidSubmit when submitting a valid form', () => {
-    let isCalled = sinon.spy();
+    const isCalled = jest.fn();
+
     class TestForm extends React.Component {
       render() {
         return (
@@ -433,14 +414,16 @@ describe('Update a form', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const FoundForm = form.find(TestForm);
     FoundForm.simulate('submit');
-    expect(isCalled.called).toEqual(true);
+    expect(isCalled).toHaveBeenCalled();
   });
 
   it('should trigger an onInvalidSubmit when submitting an invalid form', () => {
-    let isCalled = sinon.spy();
+    const isCalled = jest.fn();
+
     class TestForm extends React.Component {
       render() {
         return (
@@ -450,17 +433,19 @@ describe('Update a form', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
 
     const FoundForm = form.find(TestForm);
     FoundForm.simulate('submit');
-    expect(isCalled.called).toEqual(true);
+    expect(isCalled).toHaveBeenCalled();
   });
 });
 
 describe('value === false', () => {
   it('should call onSubmit correctly', () => {
-    const onSubmit = sinon.spy();
+    const onSubmit = jest.fn();
+
     class TestForm extends React.Component {
       render() {
         return (
@@ -474,20 +459,23 @@ describe('value === false', () => {
 
     const form = mount(<TestForm />);
     form.simulate('submit');
-    expect(onSubmit.calledWith({ foo: false })).toEqual(true);
+    expect(onSubmit).toHaveBeenCalledWith({ foo: false }, expect.any(Function), expect.any(Function));
   });
 
   it('should allow dynamic changes to false', () => {
-    const onSubmit = sinon.spy();
+    const onSubmit = jest.fn();
+
     class TestForm extends React.Component {
       state = {
         value: true,
       };
+
       changeValue() {
         this.setState({
           value: false,
         });
       }
+
       render() {
         return (
           <Formsy onSubmit={onSubmit}>
@@ -501,7 +489,7 @@ describe('value === false', () => {
     const form = mount(<TestForm />);
     form.instance().changeValue();
     form.simulate('submit');
-    expect(onSubmit.calledWith({ foo: false })).toEqual(true);
+    expect(onSubmit).toHaveBeenCalledWith({ foo: false }, expect.any(Function), expect.any(Function));
   });
 
   it('should say the form is submitted', () => {
@@ -515,6 +503,7 @@ describe('value === false', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const input = form.find(TestInput);
     expect(input.instance().isFormSubmitted()).toEqual(false);
@@ -527,11 +516,13 @@ describe('value === false', () => {
       state = {
         value: true,
       };
+
       changeValue() {
         this.setState({
           value: false,
         });
       }
+
       render() {
         return (
           <Formsy>
@@ -541,6 +532,7 @@ describe('value === false', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const input = form.find(TestInput);
     const formsyForm = form.find(Formsy);
@@ -551,16 +543,66 @@ describe('value === false', () => {
     expect(input.instance().getValue()).toEqual(true);
   });
 
+  it('should be able to set a value to components with updateInputsWithValue', () => {
+    class TestForm extends React.Component {
+      state = {
+        valueFoo: true,
+        valueBar: true,
+      };
+
+      render() {
+        return (
+          <Formsy>
+            <TestInput name="foo" value={this.state.valueFoo} type="checkbox" />
+            <TestInput name="bar" value={this.state.valueBar} type="checkbox" />
+            <button type="submit">Save</button>
+          </Formsy>
+        );
+      }
+    }
+
+    const form = mount(<TestForm />);
+    const inputs = form.find(TestInput);
+    const formsyForm = form.find(Formsy);
+    expect(
+      inputs
+        .at(0)
+        .instance()
+        .getValue(),
+    ).toEqual(true);
+    expect(
+      inputs
+        .at(1)
+        .instance()
+        .getValue(),
+    ).toEqual(true);
+    formsyForm.instance().updateInputsWithValue({ foo: false });
+    expect(
+      inputs
+        .at(0)
+        .instance()
+        .getValue(),
+    ).toEqual(false);
+    expect(
+      inputs
+        .at(1)
+        .instance()
+        .getValue(),
+    ).toEqual(true);
+  });
+
   it('should be able to reset the form using custom data', () => {
     class TestForm extends React.Component {
       state = {
         value: true,
       };
+
       changeValue() {
         this.setState({
           value: false,
         });
       }
+
       render() {
         return (
           <Formsy>
@@ -570,6 +612,7 @@ describe('value === false', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const input = form.find(TestInput);
     const formsyForm = form.find(Formsy);
@@ -596,6 +639,7 @@ describe('.reset()', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
     const input = form.find(TestInput);
     const formsyForm = form.find(Formsy);
@@ -609,18 +653,18 @@ describe('.reset()', () => {
 
 describe('.isChanged()', () => {
   it('initially returns false', () => {
-    const hasOnChanged = sinon.spy();
+    const hasOnChanged = jest.fn();
     const form = mount(
       <Formsy onChange={hasOnChanged}>
         <TestInput name="one" value="foo" />
       </Formsy>,
     );
     expect(form.instance().isChanged()).toEqual(false);
-    expect(hasOnChanged.called).toEqual(false);
+    expect(hasOnChanged).not.toHaveBeenCalled();
   });
 
   it('returns true when changed', () => {
-    const hasOnChanged = sinon.spy();
+    const hasOnChanged = jest.fn();
     const form = mount(
       <Formsy onChange={hasOnChanged}>
         <TestInput name="one" value="foo" />
@@ -631,11 +675,11 @@ describe('.isChanged()', () => {
       target: { value: 'bar' },
     });
     expect(form.instance().isChanged()).toEqual(true);
-    expect(hasOnChanged.calledWith({ one: 'bar' })).toEqual(true);
+    expect(hasOnChanged).toHaveBeenCalledWith({ one: 'bar' }, true);
   });
 
   it('returns false if changes are undone', () => {
-    const hasOnChanged = sinon.spy();
+    const hasOnChanged = jest.fn();
     const form = mount(
       <Formsy onChange={hasOnChanged}>
         <TestInput name="one" value="foo" />
@@ -645,29 +689,32 @@ describe('.isChanged()', () => {
     input.simulate('change', {
       target: { value: 'bar' },
     });
-    expect(hasOnChanged.calledWith({ one: 'bar' })).toBe(true);
+    expect(hasOnChanged).toHaveBeenCalledWith({ one: 'bar' }, true);
 
     input.simulate('change', {
       target: { value: 'foo' },
     });
+
     expect(form.instance().isChanged()).toEqual(false);
-    expect(hasOnChanged.calledWith({ one: 'foo' })).toBe(true);
+    expect(hasOnChanged).toHaveBeenCalledWith({ one: 'foo' }, false);
   });
 });
 
 describe('form valid state', () => {
   it('should allow to be changed with updateInputsWithError', () => {
+    let isValid = true;
+
     class TestForm extends React.Component {
-      state = { isValid: true };
       onValidSubmit = (model, reset, updateInputsWithError) => {
         updateInputsWithError({ foo: 'bar' }, true);
       };
       onValid = () => {
-        this.setState({ isValid: true });
+        isValid = true;
       };
       onInvalid = () => {
-        this.setState({ isValid: false });
+        isValid = false;
       };
+
       render() {
         return (
           <Formsy onInvalid={this.onInvalid} onValid={this.onValid} onValidSubmit={this.onValidSubmit}>
@@ -676,32 +723,30 @@ describe('form valid state', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
 
-    // Wait for update
-    immediate(() => {
-      expect(form.state.isValid).toEqual(true);
-      form.simulate('submit');
+    expect(isValid).toEqual(true);
+    form.simulate('submit');
 
-      // Wait for update
-      immediate(() => {
-        expect(form.state.isValid).toEqual(false);
-      });
-    });
+    expect(isValid).toEqual(false);
   });
 
   it('should be false when validationErrors is not empty', () => {
+    let isValid = true;
+
     class TestForm extends React.Component {
-      state = { validationErrors: {}, isValid: true };
+      state = { validationErrors: {} };
       setValidationErrors = empty => {
         this.setState(!empty ? { validationErrors: { foo: 'bar' } } : { validationErrors: {} });
       };
       onValid = () => {
-        this.setState({ isValid: true });
+        isValid = true;
       };
       onInvalid = () => {
-        this.setState({ isValid: false });
+        isValid = false;
       };
+
       render() {
         return (
           <Formsy onInvalid={this.onInvalid} onValid={this.onValid} validationErrors={this.state.validationErrors}>
@@ -710,32 +755,29 @@ describe('form valid state', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
 
-    // Wait for update
-    immediate(() => {
-      expect(form.state.isValid).toEqual(true);
-      form.setValidationErrors();
+    expect(isValid).toEqual(true);
+    form.instance().setValidationErrors();
 
-      // Wait for update
-      immediate(() => {
-        expect(form.state.isValid).toEqual(false);
-      });
-    });
+    expect(isValid).toEqual(false);
   });
 
   it('should be true when validationErrors is not empty and preventExternalInvalidation is true', () => {
+    let isValid = true;
     class TestForm extends React.Component {
-      state = { validationErrors: {}, isValid: true };
+      state = { validationErrors: {} };
       setValidationErrors = empty => {
         this.setState(!empty ? { validationErrors: { foo: 'bar' } } : { validationErrors: {} });
       };
       onValid = () => {
-        this.setState({ isValid: true });
+        isValid = true;
       };
       onInvalid = () => {
-        this.setState({ isValid: false });
+        isValid = false;
       };
+
       render() {
         return (
           <Formsy
@@ -749,17 +791,12 @@ describe('form valid state', () => {
         );
       }
     }
+
     const form = mount(<TestForm />);
 
-    // Wait for update
-    immediate(() => {
-      expect(form.state.isValid).toEqual(true);
-      form.setValidationErrors();
+    expect(isValid).toEqual(true);
+    form.instance().setValidationErrors();
 
-      // Wait for update
-      immediate(() => {
-        expect(form.state.isValid).toEqual(true);
-      });
-    });
+    expect(isValid).toEqual(true);
   });
 });
