@@ -1,7 +1,7 @@
 import React from 'react';
 import { mount } from 'enzyme';
 
-import Formsy from './..';
+import Formsy from '../src';
 import { InputFactory } from '../__test_utils__/TestInput';
 
 const TestInput = InputFactory({
